% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC-sumspec.R
\name{directSumSpec}
\alias{directSumSpec}
\title{read the sum spectrum direct from file}
\usage{
directSumSpec(filename)
}
\arguments{
\item{filename}{name of h5 file}
}
\value{
sum spectrum
}
\description{
read the sum spectrum direct from file
}
