% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_creation_time.R
\name{get_creation_time}
\alias{get_creation_time}
\title{get exact time information of scans}
\usage{
get_creation_time(fid)
}
\arguments{
\item{fid}{h5d file handle}
}
\value{
start time as posix_ct
}
\description{
get exact time information of scans
}
\note{
at some point the date format changed. Currently need to try several
}
\examples{
\dontrun{
get_timing("tof_file.h5")
}
}
