% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_matrix.R
\name{get_peak_matrix}
\alias{get_peak_matrix}
\title{retrieve matrix of precalculated peaks}
\usage{
get_peak_matrix(fid)
}
\arguments{
\item{fid}{h5d file handle}
}
\description{
retrieve matrix of precalculated peaks
}
