% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_scan.R
\name{get_single_scan}
\alias{get_single_scan}
\title{extract a single scan from tof spectrum}
\usage{
get_single_scan(toffile, id)
}
\arguments{
\item{toffile}{file name}

\item{id}{index of scan}
}
\value{
vector of single scan
}
\description{
extract a single scan from tof spectrum
}
\note{
the data block is opened with each individual call in scripts
it may prove usefull to open it once and make smaller interactions
with it.

this function does not check whether the scan is in the range
of measurement.  Checks will need to be done manually!

Theoretically the full matrix of scans could be loaded similar
to get_peak_matrix - but this will take loads of memory and usually
is not desirable.
}
\examples{
\dontrun{
get_single_scan("toffile.h5",1)
}
}
\seealso{
\code{get_sum_spec} which return the precalculated sum of
all individual scans
}
