% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_spec.R
\name{get_sum_spec}
\alias{get_sum_spec}
\title{retrieve sum spectrum}
\usage{
get_sum_spec(tof_file)
}
\arguments{
\item{tof_file}{h5d file handle}
}
\description{
retrieve sum spectrum
}
\examples{
\dontrun{
get_sum_spec("toffile.h5")
}
}
\seealso{
\code{get_single_scan} which returns individual scans
}
