% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timing.R
\name{get_timing}
\alias{get_timing}
\title{get exact time information of scans}
\usage{
get_timing(fid)
}
\arguments{
\item{fid}{h5d file handle}
}
\value{
vector of all measurment times
}
\description{
get exact time information of scans
}
\examples{
\dontrun{
get_timing("tof_file.h5")
}
}
