% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC-initialize.R
\docType{methods}
\name{initialize,TofClass-method}
\alias{initialize,TofClass-method}
\title{initializes TofClass based on filename}
\usage{
\S4method{initialize}{TofClass}(.Object, filename)
}
\description{
initializes TofClass based on filename
}
