% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RawHandle.R
\name{raw_handle}
\alias{raw_handle}
\title{S3 class constructor for TOF object}
\usage{
raw_handle(filename)
}
\arguments{
\item{filename}{name of the h5 file}
}
\value{
S3 object
}
\description{
stores repeatedly used object information
}
\examples{
\dontrun{
rh <- raw_handle("mytof.h5")
}
}
