% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawTof.R
\docType{package}
\name{rawTof-package}
\alias{rawTof-package}
\title{Interact with Ionicon h5 Tof files.}
\description{
Interact with Ionicon h5 Tof files.
}
\references{
\url{https://www.hdfgroup.org}
\url{http://www.ionicon.com}
\url{https://sites.google.com/site/ptrtof/home}

A package to interact directly with the h5 files produced by Ionicon
data acquisition.
}
