% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stored_masscal.R
\name{read_stored_masscal}
\alias{read_stored_masscal}
\title{reads the mass calibration from a file}
\usage{
read_stored_masscal(file)
}
\arguments{
\item{file}{location of h5 file}
}
\value{
list of conversion functions and parameters
}
\description{
reads the mass calibration from a file
}
