% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofClass.R, R/TC-scan_ind.R
\name{scan_ind}
\alias{scan_ind}
\alias{scan_ind}
\title{read the spectrum of scan i}
\usage{
scan_ind(.Object, ind)

scan_ind(.Object, ind)
}
\description{
read the spectrum of scan i

read the spectrum of scan i
}
