% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stored_mass_cal.R
\name{stored_mass_cal.tof_h5}
\alias{stored_mass_cal.tof_h5}
\title{reads mass calibration from H5 file}
\usage{
stored_mass_cal.tof_h5(fid)
}
\arguments{
\item{fid}{a h5f file handle}
}
\value{
list of functions to convert between mass and index
}
\description{
reads mass calibration from H5 file
}
\note{
This mass calibration is obtained bz TOF DAQ software.
It uses the averaged spectrum to obtain the calibration.  There is a second
mass calibration table in the file which uses averages of chunks.  The other one is just not used in this package.

This function is currently only used for convenience and uses the
simples formula, hardcoded.   I am not sure if I will reuse it or even
bother to adjust for all the available formulas.
}
