% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC-sumspec.R
\name{storedMassCal}
\alias{storedMassCal}
\title{read stored mass cal}
\usage{
storedMassCal(.Object)
}
\arguments{
\item{.Object}{an object}
}
\value{
list with mass cal functions
}
\description{
read stored mass cal
}
