% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofClass.R, R/TC-finalize.R, R/TC-scan_block.R,
%   R/TC-scan_ind.R, R/TC-sumspec.R
\docType{class}
\name{TofClass}
\alias{TofClass}
\alias{TofClass-class}
\alias{finalize}
\alias{scan_block}
\alias{scan_block,TofClass-method}
\alias{scan_ind,TofClass-method}
\alias{sumspec,TofClass-method}
\title{An S4 class to hold reused values for reading the spectrum}
\usage{
finalize(.Object)

scan_block(.Object, indrange)

\S4method{scan_block}{TofClass}(.Object, indrange)

\S4method{scan_ind}{TofClass}(.Object, ind)

\S4method{sumspec}{TofClass}(.Object)
}
\arguments{
\item{.Object}{a TofH5 object}

\item{indrange}{list with two elements where the first is the start index and the second the number of scans}

\item{ind}{as numeric index of scan}
}
\value{
matrix of counts

matrix of counts
}
\description{
An S4 class to hold reused values for reading the spectrum

do cleanup, call this before leaving

read a block with specified indices

read a block with specified indices

read the spectrum of scan i

read the sum spectrum
}
\section{Slots}{

\describe{
\item{\code{filename}}{location of original h5 file}

\item{\code{points}}{per scan}

\item{\code{nscans}}{number of scans}

\item{\code{.dims}}{internal dimensions}

\item{\code{.datafile}}{h5 handle of tof data file}

\item{\code{.scandata}}{h5 handle of scan data block}

\item{\code{.scanspace}}{h5 handle of scan data array}

\item{\code{single_ion_signal}}{cps value of single count}
}}

