% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofClass.R
\docType{class}
\name{TofClass}
\alias{TofClass}
\alias{TofClass-class}
\title{An S4 class to hold reused values for reading the spectrum}
\description{
An S4 class to hold reused values for reading the spectrum
}
\section{Slots}{

\describe{
\item{\code{filename}}{location of original h5 file}

\item{\code{points}}{per scan}

\item{\code{nscans}}{number of scans}

\item{\code{.dims}}{internal dimensions}
}}

