% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_scan.R
\name{get_single_scan}
\alias{get_single_scan}
\title{extract a single scan from tof spectrum}
\usage{
get_single_scan(fid, id)
}
\arguments{
\item{fid}{h5d file handle}

\item{id}{index of scan}
}
\value{
vector of single scan
}
\description{
Note: this function does not check whether the scan is in the range
of measurement.  Checks will need to be done manually!
Note2: Theoretically the full matrix of scans could be loaded similar
to get_peak_matrix - but this will take loads of memory and usually
is not desirable.
}
