% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_calib.R
\name{masscal_helper}
\alias{masscal_helper}
\title{create functions to work with existing mass calibration}
\usage{
masscal_helper(fid)
}
\arguments{
\item{fid}{h5d file handle}
}
\description{
create functions to work with existing mass calibration
}
\examples{
\dontrun{
fid <- h5open('toffile.h5')
mc <- masscal_helper(fid)
with(mc, to_index(c(45,47)))
with(mc, to_mass(c(2000:3000)))
}
}
