% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TC-scan_ind.R
\name{scan_ind}
\alias{scan_ind}
\title{read the spectrum of scan i}
\usage{
scan_ind(.Object, ind)
}
\arguments{
\item{.Object}{a  object}

\item{ind}{as numeric index of scan}
}
\value{
scan as list of values
}
\description{
read the spectrum of scan i
}
