% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TofClass.R, R/TC-finalize.R, R/TC-initialize.R,
%   R/TC-scan_block.R, R/TC-scan_ind.R, R/TC-sumspec.R
\docType{class}
\name{TofClass}
\alias{TofClass}
\alias{TofClass-class}
\alias{finalize}
\alias{initialize}
\alias{scan_block}
\alias{scan_ind}
\alias{sumspec}
\title{An S4 class to hold reused values for reading the spectrum}
\usage{
finalize(.Object)

initialize(.Object, filename)

\S4method{initialize}{TofClass}(.Object, filename)

scan_block(.Object, indrange)

scan_ind(.Object, ind)

sumspec(.Object)
}
\description{
An S4 class to hold reused values for reading the spectrum

do cleanup, call this before leaving

initializes TofClass based on filename

read a block with specified indices

read the spectrum of scan i

read the sum spectrum
}
\section{Slots}{

\describe{
\item{\code{filename}}{location of original h5 file}

\item{\code{points}}{per scan}

\item{\code{nscans}}{number of scans}

\item{\code{.dims}}{internal dimensions}
}}

